pyproject.toml
[project]
name = "divoom-hub"
version = "0.1.0"
description = "Pixoo 64 smart display hub: clock, bus times, and weather at a glance"
requires-python = ">=3.10"
dependencies = [
    "pixoo",
    "Pillow",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint.per-file-ignores]
# Entry point needs sys.path manipulation before src.* imports
"src/main.py" = ["E402"]

[tool.pytest.ini_options]
testpaths = ["tests"]
